[#31] refactor: 폴더 구조 변경 (Personal 페이지 내부로 이동)
