+++ front/src/App.js
@@ -1,19 +1,19 @@
 import './App.css';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
+import reset from 'theme/globalStyle';
+import { Global } from '@emotion/react';
+import AppLayout from 'components/AppLayout';
 import Main from 'pages/Main';
 import Store from 'pages/Main/Stores';
 import Categories from 'pages/Categories';
+import Category from 'pages/Categories/Category';
 import Apply from 'pages/Apply';
 import Personal from 'pages/Personal';
-import { Global } from '@emotion/react';
-import reset from 'theme/globalStyle';
-import AppLayout from 'components/AppLayout';
-import Category from 'pages/Categories/Category';
-import MyApply from 'pages/MyApply';
-import MyReview from 'pages/MyReview';
-import Qna from 'pages/Qna';
-import Question from 'pages/Question';
-import Quit from 'pages/Quit';
+import MyApply from 'pages/Personal/MyApply';
+import MyReview from 'pages/Personal/MyReview';
+import Qna from 'pages/Personal/Qna';
+import Question from 'pages/Personal/Question';
+import Quit from 'pages/Personal/Quit';
 
 function App() {
   return (
