+++ front/src/pages/Personal/Quit/index.js
