+++ front/src/pages/Personal/Quit/style.js
