--- front/src/pages/Submit/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Submit = () => {
-  return <div>Submit 페이지입니다.</div>;
-};
-
-export default Submit;
